+++ scripts/macros/2024/classFeatures/pugilist/pugilistUnarmedStrike.js
@@ -1,10 +1,10 @@
 export let pugilistUnarmedStrike = {
     name: 'Unarmed Strike (Pugilist)',
-    version: '1.4.25',
+    version: '1.5.34',
     rules: 'modern',
     ddbi: {
         restrictedItems: {
-            'Unarmed Strike': {
+            'Unarmed Strike 5': {
                 originalName: 'Unarmed Strike',
                 requiredClass: 'Pugilist',
                 requiredSubclass: null,