+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # 1.5.35 Release:
 ## Update Notes:
 - This is not a V14 update. An update for V14 is still a work in progress.
+## New Modern Content:
+- Unarmed Strike (Monk, modern)
 ## Bug Fixes:
 - Iron Chin
+- Unarmed Strike (Pugilist)
 
 # 1.5.34 Release:
 ## Update Notes: